+++ src/pages/DashboardPage/index.jsx
@@ -2,22 +2,14 @@ import { useState } from "react";
 import DateSelector from "./DateSelector";
 import MonthlyUserInfo from "./MonthlyUserInfo";
 import MonthlyUserTable from "./MonthlyUserTable";
+import MonthlyPerfectUserTable from "./MonthlyPerfectUserTable";
 import Grid from "@mui/material/Grid";
 import Paper from "@mui/material/Paper";
 import Container from "@mui/material/Container";
 import Card from "@mui/material/Card";
 import Typography from "@mui/material/Typography";
 import CardContent from "@mui/material/CardContent";
-import { DataGrid, GridToolbarContainer, GridToolbarExport } from "@mui/x-data-grid";
 import Box from "@mui/material/Box";
-
-function CustomToolbar() {
-  return (
-    <GridToolbarContainer>
-      <GridToolbarExport />
-    </GridToolbarContainer>
-  );
-}
 
 function DashboardPage() {
   return (
@@ -34,26 +26,7 @@ function DashboardPage() {
         <MonthlyUserTable />
 
         {/* EXPLAIN: 이번 달 개근자 목록 테이블 */}
-        {/* <Grid item xs={4}>
-          <Card>
-            <CardContent>
-              <Typography variant="h5" component="div">
-                이번 달 개근자 목록
-              </Typography>
-
-              <Box sx={{ mt: 1, height: 400, width: "100%" }}>
-                <DataGrid
-                  rows={rows}
-                  columns={columns}
-                  disableRowSelectionOnClick
-                  components={{
-                    Toolbar: CustomToolbar,
-                  }}
-                />
-              </Box>
-            </CardContent>
-          </Card>
-        </Grid> */}
+        <MonthlyPerfectUserTable />
 
         {/* EXPLAIN: 출석 데이터 수정 */}
         <Grid item xs={12}>
